Fixed build errors

+++ TD1/exercice_1/include/fibonacci.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <stdint.h>
 
-unsigned long long int fibonacci(unsigned int n, long long int DPFib[]);
-unsigned long long int fibonacci_naive(unsigned int n);
+uint64_t fibonacci(unsigned int n, uint64_t DPFib[]);
+uint64_t fibonacci_naive(unsigned int n);
